Separate native and zk side to different files for short header proof provider
