+++ crates/short-header-proof-provider/src/lib.rs
@@ -1,19 +1,11 @@
-use std::marker::PhantomData;
 #[cfg(feature = "native")]
-use std::sync::Arc;
+mod native;
+mod zk;
 
-use borsh::BorshDeserialize;
 #[cfg(feature = "native")]
-use futures::executor::block_on;
+pub use native::*;
 use once_cell::sync::OnceCell;
-#[cfg(feature = "native")]
-use sov_db::ledger_db::LedgerDB;
-#[cfg(feature = "native")]
-use sov_db::ledger_db::SharedLedgerOps;
-use sov_modules_api::DaSpec;
-use sov_rollup_interface::da::VerifableShortHeaderProof;
-#[cfg(feature = "native")]
-use sov_rollup_interface::services::da::DaService;
+pub use zk::*;
 
 /// Short Header Proof Provider
 /// This trait is used to get the short header proof by the l1 hash
@@ -25,62 +17,3 @@ pub trait ShortHeaderProofProvider: Send + Sync {
 
 pub static SHORT_HEADER_PROOF_PROVIDER: OnceCell<Box<dyn ShortHeaderProofProvider>> =
     OnceCell::new();
-
-#[cfg(feature = "native")]
-pub struct NativeShortHeaderProofProviderService<Da: DaService> {
-    pub da_service: Arc<Da>,
-    pub ledger_db: LedgerDB,
-}
-
-#[cfg(feature = "native")]
-impl<Da: DaService> NativeShortHeaderProofProviderService<Da> {
-    pub fn new(da_service: Arc<Da>, ledger_db: LedgerDB) -> Self {
-        Self {
-            da_service,
-            ledger_db,
-        }
-    }
-}
-
-#[cfg(feature = "native")]
-impl<Da: DaService> ShortHeaderProofProvider for NativeShortHeaderProofProviderService<Da> {
-    fn get_and_verify_short_header_proof_by_l1_hash(&self, block_hash: [u8; 32]) -> bool {
-        let block = block_on(self.da_service.get_block_by_hash(block_hash.into())).unwrap();
-        let shp = Da::block_to_short_header_proof(block);
-        self.ledger_db
-            .put_short_header_proof_by_l1_hash(
-                &block_hash,
-                borsh::to_vec(&shp).expect("Should serialize short header proof"),
-            )
-            .expect("Should save short header proof");
-        shp.verify().is_ok()
-    }
-}
-
-pub struct ZkShortHeaderProofProviderService<Da: DaSpec> {
-    short_header_proofs: Vec<([u8; 32], Vec<u8>)>,
-    phantom: PhantomData<Da>,
-}
-impl<Da: DaSpec> ZkShortHeaderProofProviderService<Da> {
-    pub fn new(short_header_proofs: Vec<([u8; 32], Vec<u8>)>) -> Self {
-        Self {
-            short_header_proofs,
-            phantom: PhantomData,
-        }
-    }
-}
-impl<Da: DaSpec> ShortHeaderProofProvider for ZkShortHeaderProofProviderService<Da> {
-    fn get_and_verify_short_header_proof_by_l1_hash(&self, block_hash: [u8; 32]) -> bool {
-        if let Some(pos) = self
-            .short_header_proofs
-            .iter()
-            .position(|(l1_hash, _)| l1_hash == &block_hash)
-        {
-            let shp =
-                Da::ShortHeaderProof::try_from_slice(&self.short_header_proofs[pos].1).unwrap();
-            return shp.verify().is_ok();
-        }
-        // If proof not found also return false
-        false
-    }
-}
